fix some warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4206 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/bsmt2000.c
@@ -503,7 +503,7 @@ int BSMT2000_sh_start(const struct MachineSound *msound)
 
 		/* initialize the regions */
 		bsmt2000[i].region_base = (INT8 *)memory_region(intf->region[i]);
-		bsmt2000[i].total_banks = memory_region_length(intf->region[i]) / 0x10000;
+		bsmt2000[i].total_banks = (int)(memory_region_length(intf->region[i]) / 0x10000);
 
 		/* init the voices */
 		init_all_voices(&bsmt2000[i]);
